# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Propify API",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/posts/{id}/comments": {
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "Store a newly created comment in storage",
                "description": "Store Comment",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "comment",
                        "in": "body",
                        "description": "Comment that should be stored",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Comment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Comment"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/requests/{id}/comments": {
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "Store a newly created comment in storage",
                "description": "Store Comment",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "comment",
                        "in": "body",
                        "description": "Comment that should be stored",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Comment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Comment"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/products/{id}/comments": {
            "post": {
                "tags": [
                    "Marketplace"
                ],
                "summary": "Store a newly created comment in storage",
                "description": "Store Comment",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "comment",
                        "in": "body",
                        "description": "Comment that should be stored",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Comment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Comment"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/comments": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Get a listing of the Comments.",
                "description": "Get all Comments",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "The id of the commentable to retrieve comments for",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "commentable",
                        "in": "query",
                        "description": "The type of the commentable to retrieve comments for",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Comment"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/comments/{id}": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Get a listing of the children comments.",
                "description": "Get children comments",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the comment to retrieve children for",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Comment"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Comment"
                ],
                "summary": "Update a created comment in storage",
                "description": "Update Comment",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "comment",
                        "in": "body",
                        "description": "Comment that should be updated",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Comment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Comment"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comment"
                ],
                "summary": "Detele a created comment in storage",
                "description": "Delete Comment",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "comment",
                        "in": "body",
                        "description": "Comment that should be deleted",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Comment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Comment"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Comment": {
            "required": [
                "comment"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "user_id": {
                    "description": "user_id",
                    "type": "integer",
                    "format": "int32"
                },
                "type": {
                    "description": "type",
                    "type": "integer",
                    "format": "int32"
                },
                "comment": {
                    "description": "comment body",
                    "type": "string"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "passport-swaggervel_auth": {
            "type": "oauth2",
            "description": "OAuth2 grant provided by Laravel Passport",
            "flow": "accessCode",
            "authorizationUrl": "/oauth/authorize",
            "tokenUrl": "/oauth/token",
            "scopes": {}
        }
    }
}